// src/modules/merchandising/reports/report/monthly-order-vs-shipment-status-report/monthly-order-vs-shipment-status-report-type.ts
export type IMonthlyOrderVsShipmentStatusReport = {
  ORDER_DATE: string | null;
  ORDER_QTY: number | null;
  ORDER_MIN: number | null;
  ORDER_FOB: number | null;
  ORDER_CM: number | null;
  SHIP_DATE: string | null;
  SHIP_QTY: number | null;
  SHIP_MIN: number | null;
  SHIP_FOB: number | null;
  SHIP_CM: number | null;
  COMPANY_NAME: string | null;
  COMPANY_ADDRESS: string | null;
  COMPANY_REMARKS: string | null;
};